fix: Place email column unique, place lombok in NinjModel and MissionsModel classes
