--- src/main/java/com/dev/myninjaregister/ninjaregister1/NinjaModel.java
@@ -1,50 +1,0 @@
-package com.dev.myninjaregister.ninjaregister1;
-
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import org.javers.core.metamodel.annotation.Entity;
-@Entity
-@Table (name="tb_ninja_register")
-public class NinjaModel {
-    @Id
-    @GeneratedValue(strategy= GenerationType.IDENTITY)
-    private Long id;
-    private String name;
-    private String email;
-    private int age;
-
-    public NinjaModel() {
-    }
-
-    public NinjaModel(String name, String email, int age) {
-        this.name = name;
-        this.email = email;
-        this.age = age;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-}
